WIP: Refactor procsignals and interrupts.
Previously, "procsignals" were implemented with shared memory flags and
a SIGUSR1 handler that would read shared memory "reason" flags an set
process-local "interrupt" flag variables, and then set the process latch
for race-free wakeup.  The next call to CHECK_FOR_INTERRUPTS() would see
the interrupt flags and perform the relevant action.

Instead, have the sender of the procsignal set the latch directly,
have CHECK_FOR_INTERRUPTS() read the "reason" flags directly from shared
memory, and get rid of all the global variables.

This is simplier, and gets away from the assumption that every backend
is a process that can be signaled.  This removes an obstacle to running
backends in threads (or, rather, punts to the problem to the latch
module).

There are several other conditions that also handled by
CHECK_FOR_INTERRUPTS(), but are not generated by other backends sending
signals, or are set by other signal handlers; these include interrupts
triggered by timers and socket events, and are made into "procsignals"
too, for consistency.  These can be set (ie by the same backend) with
ProcSignalRaise(), but it's possible that in future we might have a
reason for another backend to set them too, so it seems like a good idea
to have a single system, effectively merging the concept of
"procsignals" and "interrupts".

There are still remaining reasons to use signals: raw usage of SIGUSR1
... XXX

XXX The timeout stuff needs more work

XXX The recovery conflict stuff needs more work

XXX The new CFI needs to be inline