WIP Do not pollute Cygwin namespace with Windows headers.
Establish that __CYGWIN__ and WIN32 should generally not be defined at
the same time.  If it ever turns out ot be necessary to access Windows
APIs directly when building for Cygwin, that should be done in an
extremely localized way, not allowing Windows macros and declarations to
leak into other translation units.

dirmod.c initially looked like a potential case for a localised
exemption, but on closer inspection it doesn't currently have any reason
to include <windows.h> -- so don't.

In passing, remove anachronistic comments about ancient Windows
versions.